Move `external_tracer` from bus mapping to new crate only for test (#303)
* Move `external-tracer` from `bus-mapping` to a new crate only for test.

* Update external-tracer and mock of bus-mapping.

* Fix clippy.

* Fix failed doctest.

* Fix failed test.

* Move errors of external-tracer from bus-mapping to eth-types. And update code format.

* Move functions for generating Gethdata to a new crate mock.
