+++ bus-mapping/src/evm/opcodes/dup.rs
@@ -32,13 +32,9 @@ impl<const N: usize> Opcode for Dup<N> {
 #[cfg(test)]
 mod dup_tests {
     use super::*;
-    use crate::{
-        bytecode,
-        circuit_input_builder::{ExecStep, TransactionContext},
-        mock,
-    };
+    use crate::circuit_input_builder::{ExecStep, TransactionContext};
     use eth_types::evm_types::StackAddress;
-    use eth_types::word;
+    use eth_types::{bytecode, word};
     use pretty_assertions::assert_eq;
 
     #[test]
@@ -55,7 +51,9 @@ mod dup_tests {
         };
 
         // Get the execution steps from the external tracer
-        let block = mock::BlockData::new_single_tx_trace_code_at_start(&code).unwrap();
+        let block = crate::mock::BlockData::new_from_geth_data(
+            mock::new_single_tx_trace_code_at_start(&code).unwrap(),
+        );
 
         let mut builder = block.new_circuit_input_builder();
         builder.handle_tx(&block.eth_tx, &block.geth_trace).unwrap();
