+++ bus-mapping/src/evm/opcodes/push.rs
@@ -32,13 +32,9 @@ impl<const N: usize> Opcode for Push<N> {
 #[cfg(test)]
 mod push_tests {
     use super::*;
-    use crate::{
-        bytecode,
-        circuit_input_builder::{ExecStep, TransactionContext},
-        mock,
-    };
+    use crate::circuit_input_builder::{ExecStep, TransactionContext};
     use eth_types::evm_types::StackAddress;
-    use eth_types::word;
+    use eth_types::{bytecode, word};
     use pretty_assertions::assert_eq;
 
     #[test]
@@ -52,7 +48,9 @@ mod push_tests {
         };
 
         // Get the execution steps from the external tracer
-        let block = mock::BlockData::new_single_tx_trace_code_at_start(&code).unwrap();
+        let block = crate::mock::BlockData::new_from_geth_data(
+            mock::new_single_tx_trace_code_at_start(&code).unwrap(),
+        );
 
         let mut builder = block.new_circuit_input_builder();
         builder.handle_tx(&block.eth_tx, &block.geth_trace).unwrap();
