+++ bus-mapping/src/lib.rs
@@ -49,14 +49,13 @@
 //!
 //! ```rust
 //! use bus_mapping::Error;
-//! use bus_mapping::mock;
 //! use bus_mapping::state_db::{self, StateDB, CodeDB};
 //! use eth_types::{
 //!     self, Address, Word, Hash, U64, GethExecTrace, GethExecStep, ChainConstants
 //! };
 //! use eth_types::evm_types::Gas;
+//! use eth_types::geth_types::BlockConstants;
 //! use bus_mapping::circuit_input_builder::CircuitInputBuilder;
-//! use bus_mapping::external_tracer::BlockConstants;
 //! use pairing::arithmetic::FieldExt;
 //!
 //! let input_trace = r#"
@@ -218,16 +217,13 @@
 #![allow(clippy::upper_case_acronyms)] // Too pedantic
 
 extern crate alloc;
+pub mod circuit_input_builder;
 pub mod error;
 pub mod evm;
 pub mod exec_trace;
-pub mod external_tracer;
-pub mod operation;
-#[macro_use]
-pub mod bytecode;
-pub mod circuit_input_builder;
 pub(crate) mod geth_errors;
 pub mod mock;
+pub mod operation;
 pub mod rpc;
 pub mod state_db;
 pub use error::Error;
