+++ external-tracer/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "external-tracer"
+version = "0.1.0"
+edition = "2021"
+authors = ["The appliedzkp team"]
+
+[dependencies]
+eth-types = { path = "../eth-types" }
+geth-utils = { path = "../geth-utils" }
+serde = {version = "1.0.130", features = ["derive"] }
+serde_json = "1.0.66"
