+++ zkevm-circuits/src/evm_circuit/execution/pop.rs
@@ -75,8 +75,7 @@ impl<F: FieldExt> ExecutionGadget<F> for PopGadget<F> {
 #[cfg(test)]
 mod test {
     use crate::{evm_circuit::test::rand_word, test_util::run_test_circuits};
-    use bus_mapping::bytecode;
-    use eth_types::Word;
+    use eth_types::{bytecode, Word};
 
     fn test_ok(value: Word) {
         let bytecode = bytecode! {
