+++ zkevm-circuits/src/evm_circuit/execution/signed_comparator.rs
@@ -197,9 +197,8 @@ impl<F: FieldExt> ExecutionGadget<F> for SignedComparatorGadget<F> {
 
 #[cfg(test)]
 mod test {
-    use bus_mapping::bytecode;
     use eth_types::evm_types::OpcodeId;
-    use eth_types::Word;
+    use eth_types::{bytecode, Word};
 
     use crate::{evm_circuit::test::rand_word, test_util::run_test_circuits};
