+++ zkevm-circuits/src/state_circuit/state.rs
@@ -1249,11 +1249,9 @@ impl<
 #[cfg(test)]
 mod tests {
     use super::*;
-    use bus_mapping::bytecode;
-    use bus_mapping::mock;
     use bus_mapping::operation::{MemoryOp, Operation, RWCounter, StackOp, StorageOp, RW};
     use eth_types::evm_types::{MemoryAddress, StackAddress};
-    use eth_types::{address, Word};
+    use eth_types::{address, bytecode, Word};
     use halo2::arithmetic::BaseExt;
     use halo2::dev::{MockProver, VerifyFailure::ConstraintNotSatisfied, VerifyFailure::Lookup};
     use pairing::bn256::Fr;
@@ -1946,7 +1944,9 @@ mod tests {
             MLOAD
             STOP
         };
-        let block = mock::BlockData::new_single_tx_trace_code_at_start(&bytecode).unwrap();
+        let block = bus_mapping::mock::BlockData::new_from_geth_data(
+            mock::new_single_tx_trace_code_at_start(&bytecode).unwrap(),
+        );
         let mut builder = block.new_circuit_input_builder();
         builder.handle_tx(&block.eth_tx, &block.geth_trace).unwrap();
